pyproject.toml
[project]
name = "hsa-receipt-system"
version = "0.1.0"
description = "Automated HSA receipt organization system with vision LLM extraction"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "ming"}]
license = { text = "MIT" }

dependencies = [
    # Google APIs
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.150.0",
    "gspread>=6.1.0",

    # LLM client (OpenAI-compatible API for Ollama/vLLM)
    "openai>=1.50.0",

    # Email
    "imapclient>=3.0.0",
    "email-validator>=2.2.0",

    # Image/PDF handling
    "Pillow>=11.0.0",
    "pdf2image>=1.17.0",
    "pypdf>=5.1.0",

    # Web scraping for portals (future)
    "playwright>=1.49.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",

    # Utilities
    "pyyaml>=6.0",
    "python-dateutil>=2.9.0",
    "rich>=13.9.0",
    "click>=8.1.0",
    "watchdog>=6.0.0",

    # Data handling
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
hsa = "src.pipeline:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
